pyproject.toml
[project]
name = "idinsights"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "convex>=0.7.0",
    "fastapi>=0.128.0",
    "google-cloud-documentai>=3.9.0",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
